\name{RandomLetterGenerator-package}
\alias{RandomLetterGenerator-package}
\alias{RandomLetterGenerator}
\docType{package}
\title{
\packageTitle{RandomLetterGenerator}
}
\description{
\packageDescription{RandomLetterGenerator}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RandomLetterGenerator}
\packageIndices{RandomLetterGenerator}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RandomLetterGenerator}

Maintainer: \packageMaintainer{RandomLetterGenerator}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
